fix: 親スコープ候補選択時のフォーカス確定を修正

+++ addons/katamusubi/editor/inspector/parent_scope_id_editor_property.gd
@@ -10,6 +10,7 @@ var _line_edit: LineEdit
 var _item_list: ItemList
 
 var _index: ScopeIndex
+var _last_committed_value: Variant = null
 
 
 func _init(init_index: ScopeIndex) -> void:
@@ -38,6 +39,7 @@ func _init(init_index: ScopeIndex) -> void:
 func _update_property() -> void:
 	var value := get_edited_object().get(get_edited_property()) as StringName
 	_line_edit.text = value
+	_last_committed_value = value
 
 
 func _on_text_changed(value: String) -> void:
@@ -63,14 +65,27 @@ func _on_text_submitted(text: String) -> void:
 
 
 func _on_focus_exited() -> void:
+	_complete_focus_exit.call_deferred()
+
+
+func _complete_focus_exit() -> void:
+	var focus_owner := get_viewport().gui_get_focus_owner()
+	if focus_owner != null and (
+			focus_owner == _item_list or _item_list.is_ancestor_of(focus_owner)
+	):
+		return
+
 	_commit(StringName(_line_edit.text))
 	_delete_item_list_menu()
 
 
 func _on_item_selected(index: int) -> void:
 	var key := _item_list.get_item_text(index)
+	var value := _candidates[key]
 
-	emit_changed(get_edited_property(), _candidates[key])
+	_commit(value)
+	if _line_edit.text != value:
+		_line_edit.text = value
 
 	_delete_item_list_menu()
 
@@ -82,9 +97,13 @@ func _delete_item_list_menu() -> void:
 
 ## 変更を確定する。
 func _commit(value: StringName) -> void:
+	if _last_committed_value == value:
+		return
 	if get_edited_object().get(get_edited_property()) == value:
+		_last_committed_value = value
 		return
 
+	_last_committed_value = value
 	emit_changed(get_edited_property(), value)
